requirements.txt
fastapi[all]==0.115.5
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
pymysql==1.1.1
python-dotenv==1.0.1
xgboost
requests
cryptography